Reorganize dtslint tests and upgrade npm dev dependencies
