+++ package.json
@@ -34,9 +34,8 @@
         "pack": "run-p clean:pack build && npm pack",
         "jest": "jest",
         "jest:coverage": "npm run clean:coverage && jest --coverage",
-        "dtslint:v2_8_plus": "dtslint type_tests/v2_8_plus",
         "dtslint:v2_9_plus": "dtslint type_tests/v2_9_plus",
-        "dtslint": "run-s clean:dist build:types dtslint:v2_8_plus dtslint:v2_9_plus",
+        "dtslint": "run-s clean:dist build:types dtslint:v2_9_plus",
         "test": "run-s compile prettier:test lint dtslint jest",
         "test:coverage": "run-s compile prettier:test lint dtslint jest:coverage",
         "build:travis": "run-p test:coverage build && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
@@ -48,21 +47,21 @@
     },
     "license": "MIT",
     "devDependencies": {
-        "@types/jest": "23.3.3",
-        "@types/node": "10.11.4",
+        "@types/jest": "23.3.13",
+        "@types/node": "10.12.18",
         "coveralls": "3.0.2",
-        "dtslint": "0.3.0",
+        "dtslint": "0.4.2",
         "jest": "23.6.0",
-        "npm-run-all": "4.1.3",
-        "prettier": "1.14.3",
-        "rimraf": "2.6.2",
-        "ts-jest": "23.10.3",
-        "tslint": "5.11.0",
-        "tslint-config-prettier": "1.15.0",
-        "typescript": "3.1.1"
+        "npm-run-all": "4.1.5",
+        "prettier": "1.16.1",
+        "rimraf": "2.6.3",
+        "ts-jest": "23.10.5",
+        "tslint": "5.12.1",
+        "tslint-config-prettier": "1.17.0",
+        "typescript": "3.3.3"
     },
     "peerDependencies": {
-        "typescript": ">= 2.8.1"
+        "typescript": ">= 2.9.1"
     },
     "keywords": [
         "typescript",
