--- type_tests/v2_8_plus/README.md
@@ -1,3 +1,0 @@
-### Compile-Time Tests: TypeScript 2.8+
-
-Due to breaking changes to the `keyof` operator in TypeScript 2.9, there is no way for the types of this project to be defined in a way that is both compatible with 2.9+ and with versions prior to 2.8.
