--- type_tests/v2_8_plus/index.d.ts
@@ -1,3 +1,0 @@
-// Dummy index.d.ts file for running dtslint.
-// We're not actually linting a .d.ts file; only using dtslint to verify compile time types of ts-enum-util.
-// TypeScript Version: 2.8
