--- type_tests/v2_8_plus/tsconfig.json
@@ -1,24 +1,0 @@
-/* Config used for dtslint unit tests. */
-{
-    "compilerOptions": {
-        /* Basic Options */
-        "target": "es5",
-        "module": "commonjs",
-
-        "lib": ["es6"],
-
-        "downlevelIteration": true,
-
-        /* Strict Type-Checking Options */
-        "strict": true,
-        "alwaysStrict": true,
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": true,
-
-        /* Module Resolution Options */
-        "moduleResolution": "node",
-        "baseUrl": "./"
-    }
-}
