--- type_tests/v2_8_plus/tslint.json
@@ -1,9 +1,0 @@
-{
-    "extends": "dtslint/dtslint.json",
-    "rules": {
-        "no-useless-files": false,
-        "file-name-casing": false,
-        "no-relative-import-in-test": false,
-        "array-type": false
-    }
-}
