+++ type_tests/v2_9_plus/EnumWrapper/dictionary-string.dtslint.ts
@@ -1,7 +1,7 @@
-import { $enum } from "../../../dist/types";
+import { $enum } from "ts-enum-util";
 
-// Dictionary object with number values
-declare const TestEnum: { [key: string]: number };
+// Dictionary object with string values
+declare const TestEnum: { [key: string]: string };
 
 declare const str: string;
 declare const strOrNull: string | null;
@@ -13,7 +13,7 @@ declare const numOrUndefined: number | undefined;
 
 const enumWrapper = $enum(TestEnum);
 
-// $ExpectType EnumWrapper<number, { [key: string]: number; }>
+// $ExpectType EnumWrapper<string, { [key: string]: string; }>
 enumWrapper;
 
 // $ExpectType number
@@ -31,14 +31,14 @@ enumWrapper.size = 0; // immutable
 // Also cannot test for immutability of items within the entry tuple because of
 // this change.
 // see: https://github.com/Microsoft/TypeScript/issues/26864
-const testEntry: Readonly<[string, number]> = enumWrapper[0];
+const testEntry: Readonly<[string, string]> = enumWrapper[0];
 // $ExpectError
 enumWrapper[0] = ["A", TestEnum.A]; // immutable
 
 // $ExpectType IterableIterator<string>
 enumWrapper.keys();
 
-// $ExpectType IterableIterator<number>
+// $ExpectType IterableIterator<string>
 enumWrapper.values();
 
 // NOTE: Must test via assignability rather than ExpectType because of a change
@@ -47,19 +47,19 @@ enumWrapper.values();
 // because of this change.
 // see: https://github.com/Microsoft/TypeScript/issues/26864
 const testEntryIterator: IterableIterator<
-    Readonly<[string, number]>
+    Readonly<[string, string]>
 > = enumWrapper.entries();
 for (const entry of enumWrapper.entries()) {
-    const testIteratedEntry: Readonly<[string, number]> = entry;
+    const testIteratedEntry: Readonly<[string, string]> = entry;
 }
 
 // $ExpectType void
 enumWrapper.forEach((value, key, collection, index) => {
-    // $ExpectType number
+    // $ExpectType string
     value;
     // $ExpectType string
     key;
-    // $ExpectType EnumWrapper<number, { [key: string]: number; }>
+    // $ExpectType EnumWrapper<string, { [key: string]: string; }>
     collection;
     // $ExpectType number
     index;
@@ -69,11 +69,11 @@ enumWrapper.forEach((value, key, collection, index) => {
 
 // $ExpectType number[]
 enumWrapper.map((value, key, collection, index) => {
-    // $ExpectType number
+    // $ExpectType string
     value;
     // $ExpectType string
     key;
-    // $ExpectType EnumWrapper<number, { [key: string]: number; }>
+    // $ExpectType EnumWrapper<string, { [key: string]: string; }>
     collection;
     // $ExpectType number
     index;
@@ -84,7 +84,7 @@ enumWrapper.map((value, key, collection, index) => {
 // $ExpectType string[]
 enumWrapper.getKeys();
 
-// $ExpectType number[]
+// $ExpectType string[]
 enumWrapper.getValues();
 
 // NOTE: Must test via assignability rather than ExpectType because of a change
@@ -92,7 +92,7 @@ enumWrapper.getValues();
 // Also cannot test for immutability of items within the entry tuple because of
 // this change.
 // see: https://github.com/Microsoft/TypeScript/issues/26864
-const testEntries: Readonly<[string, number]>[] = enumWrapper.getEntries();
+const testEntries: Readonly<[string, string]>[] = enumWrapper.getEntries();
 
 // $ExpectType boolean
 enumWrapper.isKey(str);
@@ -137,98 +137,98 @@ enumWrapper.asKeyOrDefault(str, str);
 enumWrapper.asKeyOrDefault(str, strOrUndefined);
 
 // $ExpectType boolean
-enumWrapper.isValue(num);
-// $ExpectType boolean
-enumWrapper.isValue(numOrNull);
-// $ExpectType boolean
-enumWrapper.isValue(numOrUndefined);
-// $ExpectError
 enumWrapper.isValue(str);
-
-if (enumWrapper.isValue(num)) {
-    // $ExpectType number
-    num;
-}
-
-if (enumWrapper.isValue(numOrNull)) {
-    // $ExpectType number
-    numOrNull;
-}
-
-if (enumWrapper.isValue(numOrUndefined)) {
-    // $ExpectType number
-    numOrUndefined;
-}
-
-// $ExpectType number
-enumWrapper.asValueOrThrow(num);
-// $ExpectType number
-enumWrapper.asValueOrThrow(numOrNull);
-// $ExpectType number
-enumWrapper.asValueOrThrow(numOrUndefined);
+// $ExpectType boolean
+enumWrapper.isValue(strOrNull);
+// $ExpectType boolean
+enumWrapper.isValue(strOrUndefined);
 // $ExpectError
+enumWrapper.isValue(num);
+
+if (enumWrapper.isValue(str)) {
+    // $ExpectType string
+    str;
+}
+
+if (enumWrapper.isValue(strOrNull)) {
+    // $ExpectType string
+    strOrNull;
+}
+
+if (enumWrapper.isValue(strOrUndefined)) {
+    // $ExpectType string
+    strOrUndefined;
+}
+
+// $ExpectType string
 enumWrapper.asValueOrThrow(str);
-
-// $ExpectType number | undefined
-enumWrapper.asValueOrDefault(num);
-// $ExpectType number | undefined
-enumWrapper.asValueOrDefault(numOrNull);
-// $ExpectType number | undefined
-enumWrapper.asValueOrDefault(numOrUndefined);
+// $ExpectType string
+enumWrapper.asValueOrThrow(strOrNull);
+// $ExpectType string
+enumWrapper.asValueOrThrow(strOrUndefined);
 // $ExpectError
+enumWrapper.asValueOrThrow(num);
+
+// $ExpectType string | undefined
 enumWrapper.asValueOrDefault(str);
-
-// $ExpectType number | undefined
-enumWrapper.asValueOrDefault(num, undefined);
-// $ExpectType number
-enumWrapper.asValueOrDefault(num, num);
-// $ExpectType number | undefined
-enumWrapper.asValueOrDefault(num, numOrUndefined);
+// $ExpectType string | undefined
+enumWrapper.asValueOrDefault(strOrNull);
+// $ExpectType string | undefined
+enumWrapper.asValueOrDefault(strOrUndefined);
 // $ExpectError
-enumWrapper.asValueOrDefault(num, str);
+enumWrapper.asValueOrDefault(num);
+
+// $ExpectType string | undefined
+enumWrapper.asValueOrDefault(str, undefined);
+// $ExpectType string
+enumWrapper.asValueOrDefault(str, str);
+// $ExpectType string | undefined
+enumWrapper.asValueOrDefault(str, strOrUndefined);
+// $ExpectError
+enumWrapper.asValueOrDefault(str, num);
 
 // $ExpectType string
-enumWrapper.getKeyOrThrow(num);
-// $ExpectType string
-enumWrapper.getKeyOrThrow(numOrNull);
-// $ExpectType string
-enumWrapper.getKeyOrThrow(numOrUndefined);
-// $ExpectError
 enumWrapper.getKeyOrThrow(str);
+// $ExpectType string
+enumWrapper.getKeyOrThrow(strOrNull);
+// $ExpectType string
+enumWrapper.getKeyOrThrow(strOrUndefined);
+// $ExpectError
+enumWrapper.getKeyOrThrow(num);
 
 // $ExpectType string | undefined
-enumWrapper.getKeyOrDefault(num);
-// $ExpectType string | undefined
-enumWrapper.getKeyOrDefault(numOrNull);
-// $ExpectType string | undefined
-enumWrapper.getKeyOrDefault(numOrUndefined);
-// $ExpectError
 enumWrapper.getKeyOrDefault(str);
+// $ExpectType string | undefined
+enumWrapper.getKeyOrDefault(strOrNull);
+// $ExpectType string | undefined
+enumWrapper.getKeyOrDefault(strOrUndefined);
+// $ExpectError
+enumWrapper.getKeyOrDefault(num);
 
 // $ExpectType string
-enumWrapper.getKeyOrDefault(num, str);
+enumWrapper.getKeyOrDefault(str, str);
 // $ExpectType string | undefined
-enumWrapper.getKeyOrDefault(num, strOrUndefined);
+enumWrapper.getKeyOrDefault(str, strOrUndefined);
 
-// $ExpectType number
+// $ExpectType string
 enumWrapper.getValueOrThrow(str);
-// $ExpectType number
+// $ExpectType string
 enumWrapper.getValueOrThrow(strOrNull);
-// $ExpectType number
+// $ExpectType string
 enumWrapper.getValueOrThrow(strOrUndefined);
 
-// $ExpectType number | undefined
+// $ExpectType string | undefined
 enumWrapper.getValueOrDefault(str);
-// $ExpectType number | undefined
+// $ExpectType string | undefined
 enumWrapper.getValueOrDefault(strOrNull);
-// $ExpectType number | undefined
+// $ExpectType string | undefined
 enumWrapper.getValueOrDefault(strOrUndefined);
 
-// $ExpectType number | undefined
+// $ExpectType string | undefined
 enumWrapper.getValueOrDefault(str, undefined);
-// $ExpectType number
-enumWrapper.getValueOrDefault(str, num);
-// $ExpectType number | undefined
-enumWrapper.getValueOrDefault(str, numOrUndefined);
-// $ExpectError
+// $ExpectType string
 enumWrapper.getValueOrDefault(str, str);
+// $ExpectType string | undefined
+enumWrapper.getValueOrDefault(str, strOrUndefined);
+// $ExpectError
+enumWrapper.getValueOrDefault(str, num);
