+++ type_tests/v2_9_plus/EnumWrapper/object-mixed.dtslint.ts
@@ -1,7 +1,11 @@
-import { $enum } from "../../../dist/types";
+import { $enum } from "ts-enum-util";
 
-// Dictionary object with a mix of number and string values
-declare const TestEnum: { [key: string]: string | number };
+// Enum-like object with mix of number and string values
+const TestEnum = {
+    A: 1,
+    B: 2,
+    C: "c"
+};
 
 declare const str: string;
 declare const strOrNull: string | null;
@@ -15,9 +19,13 @@ declare const numstr: number | string;
 declare const numstrOrNull: number | string | null;
 declare const numstrOrUndefined: number | string | undefined;
 
+declare const key: keyof typeof TestEnum;
+declare const keyOrNull: keyof typeof TestEnum | null;
+declare const keyOrUndefined: keyof typeof TestEnum | undefined;
+
 const enumWrapper = $enum(TestEnum);
 
-// $ExpectType EnumWrapper<string | number, { [key: string]: string | number; }>
+// $ExpectType EnumWrapper<string | number, { A: number; B: number; C: string; }>
 enumWrapper;
 
 // $ExpectType number
@@ -35,11 +43,11 @@ enumWrapper.size = 0; // immutable
 // Also cannot test for immutability of items within the entry tuple because of
 // this change.
 // see: https://github.com/Microsoft/TypeScript/issues/26864
-const testEntry: Readonly<[string, string | number]> = enumWrapper[0];
+const testEntry: Readonly<["A" | "B" | "C", string | number]> = enumWrapper[0];
 // $ExpectError
 enumWrapper[0] = ["A", TestEnum.A]; // immutable
 
-// $ExpectType IterableIterator<string>
+// $ExpectType IterableIterator<"A" | "B" | "C">
 enumWrapper.keys();
 
 // $ExpectType IterableIterator<string | number>
@@ -51,19 +59,21 @@ enumWrapper.values();
 // because of this change.
 // see: https://github.com/Microsoft/TypeScript/issues/26864
 const testEntryIterator: IterableIterator<
-    Readonly<[string, string | number]>
+    Readonly<["A" | "B" | "C", string | number]>
 > = enumWrapper.entries();
 for (const entry of enumWrapper.entries()) {
-    const testIteratedEntry: Readonly<[string, string | number]> = entry;
+    const testIteratedEntry: Readonly<
+        ["A" | "B" | "C", string | number]
+    > = entry;
 }
 
 // $ExpectType void
 enumWrapper.forEach((value, key, collection, index) => {
     // $ExpectType string | number
     value;
-    // $ExpectType string
+    // $ExpectType "A" | "B" | "C"
     key;
-    // $ExpectType EnumWrapper<string | number, { [key: string]: string | number; }>
+    // $ExpectType EnumWrapper<string | number, { A: number; B: number; C: string; }>
     collection;
     // $ExpectType number
     index;
@@ -75,9 +85,9 @@ enumWrapper.forEach((value, key, collection, index) => {
 enumWrapper.map((value, key, collection, index) => {
     // $ExpectType string | number
     value;
-    // $ExpectType string
+    // $ExpectType "A" | "B" | "C"
     key;
-    // $ExpectType EnumWrapper<string | number, { [key: string]: string | number; }>
+    // $ExpectType EnumWrapper<string | number, { A: number; B: number; C: string; }>
     collection;
     // $ExpectType number
     index;
@@ -85,7 +95,7 @@ enumWrapper.map((value, key, collection, index) => {
     return num;
 });
 
-// $ExpectType string[]
+// $ExpectType ("A" | "B" | "C")[]
 enumWrapper.getKeys();
 
 // $ExpectType (string | number)[]
@@ -97,7 +107,7 @@ enumWrapper.getValues();
 // this change.
 // see: https://github.com/Microsoft/TypeScript/issues/26864
 const testEntries: Readonly<
-    [string, string | number]
+    ["A" | "B" | "C", string | number]
 >[] = enumWrapper.getEntries();
 
 // $ExpectType boolean
@@ -108,35 +118,39 @@ enumWrapper.isKey(strOrNull);
 enumWrapper.isKey(strOrUndefined);
 
 if (enumWrapper.isKey(str)) {
-    // $ExpectType string
+    // $ExpectType "A" | "B" | "C"
     str;
 }
 
 if (enumWrapper.isKey(strOrNull)) {
-    // $ExpectType string
+    // $ExpectType "A" | "B" | "C"
     strOrNull;
 }
 
 if (enumWrapper.isKey(strOrUndefined)) {
-    // $ExpectType string
+    // $ExpectType "A" | "B" | "C"
     strOrUndefined;
 }
 
-// $ExpectType string
+// $ExpectType "A" | "B" | "C"
 enumWrapper.asKeyOrThrow(str);
-// $ExpectType string
+// $ExpectType "A" | "B" | "C"
 enumWrapper.asKeyOrThrow(strOrNull);
-// $ExpectType string
+// $ExpectType "A" | "B" | "C"
 enumWrapper.asKeyOrThrow(strOrUndefined);
 
-// $ExpectType string | undefined
+// $ExpectType "A" | "B" | "C" | undefined
 enumWrapper.asKeyOrDefault(str);
-// $ExpectType string | undefined
+// $ExpectType "A" | "B" | "C" | undefined
 enumWrapper.asKeyOrDefault(strOrNull);
-// $ExpectType string | undefined
+// $ExpectType "A" | "B" | "C" | undefined
 enumWrapper.asKeyOrDefault(strOrUndefined);
-// $ExpectType string | undefined
+// $ExpectType "A" | "B" | "C" | undefined
 enumWrapper.asKeyOrDefault(str, undefined);
+// $ExpectType "A" | "B" | "C"
+enumWrapper.asKeyOrDefault(str, key);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.asKeyOrDefault(str, keyOrUndefined);
 // $ExpectType string
 enumWrapper.asKeyOrDefault(str, str);
 // $ExpectType string | undefined
@@ -189,25 +203,35 @@ enumWrapper.asValueOrDefault(numstr, numstr);
 // $ExpectType string | number | undefined
 enumWrapper.asValueOrDefault(num, numstrOrUndefined);
 
-// $ExpectType string
+// $ExpectType "A" | "B" | "C"
 enumWrapper.getKeyOrThrow(numstr);
-// $ExpectType string
+// $ExpectType "A" | "B" | "C"
 enumWrapper.getKeyOrThrow(numstrOrNull);
-// $ExpectType string
+// $ExpectType "A" | "B" | "C"
 enumWrapper.getKeyOrThrow(numstrOrUndefined);
 
-// $ExpectType string | undefined
+// $ExpectType "A" | "B" | "C" | undefined
 enumWrapper.getKeyOrDefault(numstr);
-// $ExpectType string | undefined
+// $ExpectType "A" | "B" | "C" | undefined
 enumWrapper.getKeyOrDefault(numstrOrNull);
-// $ExpectType string | undefined
+// $ExpectType "A" | "B" | "C" | undefined
 enumWrapper.getKeyOrDefault(numstrOrUndefined);
 
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrDefault(numstr, key);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(numstr, keyOrUndefined);
 // $ExpectType string
 enumWrapper.getKeyOrDefault(numstr, str);
 // $ExpectType string | undefined
 enumWrapper.getKeyOrDefault(numstr, strOrUndefined);
 
+// $ExpectType string | number
+enumWrapper.getValueOrThrow(key);
+// $ExpectType string | number
+enumWrapper.getValueOrThrow(keyOrNull);
+// $ExpectType string | number
+enumWrapper.getValueOrThrow(keyOrUndefined);
 // $ExpectType string | number
 enumWrapper.getValueOrThrow(str);
 // $ExpectType string | number
