+++ type_tests/v2_9_plus/README.md
@@ -1,5 +1,4 @@
 ### Compile-Time Tests: TypeScript 2.9+
 
-Proper support for defining types with numeric enum/literal key types is only
-supported in TypeScript 2.9+, so these tests focus on using `$enum.visitValue`
-and `$enum.mapValue` with number and mixed number+string enums/literals.
+Proper support for defining types with numeric literal key types is only
+supported in TypeScript 2.9+
