+++ type_tests/v2_9_plus/mapValue/string-enum-null.dtslint.ts
@@ -1,4 +1,4 @@
-import { $enum } from "../../../dist/types";
+import { $enum } from "ts-enum-util";
 
 enum RGB {
     R = "r",
