+++ type_tests/v2_9_plus/mapValue/string-enum-undefined.dtslint.ts
@@ -1,44 +1,44 @@
-import { $enum } from "../../../dist/types";
+import { $enum } from "ts-enum-util";
 
-type RGB = "r" | "g" | "b";
+enum RGB {
+    R = "r",
+    G = "g",
+    B = "b"
+}
 
-declare const rgb: RGB | null | undefined;
+declare const rgb: RGB | undefined;
 
 // Return type is inferred
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    r: 10,
-    g: 20,
-    b: 30,
-    [$enum.handleNull]: -1,
+    [RGB.R]: 10,
+    [RGB.G]: 20,
+    [RGB.B]: 30,
     [$enum.handleUndefined]: -1
 });
 // $ExpectType string
 $enum.mapValue(rgb).with({
-    r: "10",
-    g: "20",
-    b: "30",
-    [$enum.handleNull]: "-1",
+    [RGB.R]: "10",
+    [RGB.G]: "20",
+    [RGB.B]: "30",
     [$enum.handleUndefined]: "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    r: 10,
-    g: $enum.unhandled,
-    b: 30,
-    [$enum.handleNull]: -1,
+    [RGB.R]: 10,
+    [RGB.G]: $enum.unhandled,
+    [RGB.B]: 30,
     [$enum.handleUndefined]: -1
 });
 
 // handleUnexpected is allowed
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    r: 10,
-    g: 20,
-    b: 30,
-    [$enum.handleNull]: -1,
+    [RGB.R]: 10,
+    [RGB.G]: 20,
+    [RGB.B]: 30,
     [$enum.handleUndefined]: -1,
     [$enum.handleUnexpected]: -1
 });
@@ -46,10 +46,9 @@ $enum.mapValue(rgb).with({
 // special handlers can be unhandled
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    r: 10,
-    g: 20,
-    b: 30,
-    [$enum.handleNull]: $enum.unhandled,
+    [RGB.R]: 10,
+    [RGB.G]: 20,
+    [RGB.B]: 30,
     [$enum.handleUndefined]: $enum.unhandled,
     [$enum.handleUnexpected]: $enum.unhandled
 });
@@ -57,37 +56,35 @@ $enum.mapValue(rgb).with({
 // Missing value handler causes error
 // $ExpectError
 $enum.mapValue(rgb).with({
-    r: 10,
-    b: 30,
-    [$enum.handleNull]: -1,
+    [RGB.R]: 10,
+    [RGB.B]: 30,
     [$enum.handleUndefined]: -1
 });
 
 // Unexpected value handler causes error
 $enum.mapValue(rgb).with({
-    r: 10,
+    [RGB.R]: 10,
     // $ExpectError
     oops: 42,
-    g: 20,
-    b: 30,
-    [$enum.handleNull]: -1,
-    [$enum.handleUndefined]: -1
-});
-
-// missing null handler causes error
-// $ExpectError
-$enum.mapValue(rgb).with({
-    r: 10,
-    g: 20,
-    b: 30,
+    [RGB.G]: 20,
+    [RGB.B]: 30,
     [$enum.handleUndefined]: -1
 });
 
 // missing undefined handler causes error
 // $ExpectError
 $enum.mapValue(rgb).with({
-    r: 10,
-    g: 20,
-    b: 30,
-    [$enum.handleNull]: -1
+    [RGB.R]: 10,
+    [RGB.G]: 20,
+    [RGB.B]: 30
+});
+
+// Unnecessary null handler causes error
+$enum.mapValue(rgb).with({
+    [RGB.R]: 10,
+    [RGB.G]: 20,
+    [RGB.B]: 30,
+    // $ExpectError
+    [$enum.handleNull]: -1,
+    [$enum.handleUndefined]: -1
 });
