+++ type_tests/v2_9_plus/mapValue/string-enum.dtslint.ts
@@ -1,4 +1,4 @@
-import { $enum } from "../../../dist/types";
+import { $enum } from "ts-enum-util";
 
 enum RGB {
     R = "r",
@@ -6,22 +6,20 @@ enum RGB {
     B = "b"
 }
 
-declare const rgb: RGB | undefined;
+declare const rgb: RGB;
 
 // Return type is inferred
 // $ExpectType number
 $enum.mapValue(rgb).with({
     [RGB.R]: 10,
     [RGB.G]: 20,
-    [RGB.B]: 30,
-    [$enum.handleUndefined]: -1
+    [RGB.B]: 30
 });
 // $ExpectType string
 $enum.mapValue(rgb).with({
     [RGB.R]: "10",
     [RGB.G]: "20",
-    [RGB.B]: "30",
-    [$enum.handleUndefined]: "-1"
+    [RGB.B]: "30"
 });
 
 // Return type is inferred when "unhandled" entries exist
@@ -29,8 +27,7 @@ $enum.mapValue(rgb).with({
 $enum.mapValue(rgb).with({
     [RGB.R]: 10,
     [RGB.G]: $enum.unhandled,
-    [RGB.B]: 30,
-    [$enum.handleUndefined]: -1
+    [RGB.B]: 30
 });
 
 // handleUnexpected is allowed
@@ -39,7 +36,6 @@ $enum.mapValue(rgb).with({
     [RGB.R]: 10,
     [RGB.G]: 20,
     [RGB.B]: 30,
-    [$enum.handleUndefined]: -1,
     [$enum.handleUnexpected]: -1
 });
 
@@ -49,7 +45,6 @@ $enum.mapValue(rgb).with({
     [RGB.R]: 10,
     [RGB.G]: 20,
     [RGB.B]: 30,
-    [$enum.handleUndefined]: $enum.unhandled,
     [$enum.handleUnexpected]: $enum.unhandled
 });
 
@@ -57,8 +52,7 @@ $enum.mapValue(rgb).with({
 // $ExpectError
 $enum.mapValue(rgb).with({
     [RGB.R]: 10,
-    [RGB.B]: 30,
-    [$enum.handleUndefined]: -1
+    [RGB.B]: 30
 });
 
 // Unexpected value handler causes error
@@ -66,15 +60,6 @@ $enum.mapValue(rgb).with({
     [RGB.R]: 10,
     // $ExpectError
     oops: 42,
-    [RGB.G]: 20,
-    [RGB.B]: 30,
-    [$enum.handleUndefined]: -1
-});
-
-// missing undefined handler causes error
-// $ExpectError
-$enum.mapValue(rgb).with({
-    [RGB.R]: 10,
     [RGB.G]: 20,
     [RGB.B]: 30
 });
@@ -85,6 +70,22 @@ $enum.mapValue(rgb).with({
     [RGB.G]: 20,
     [RGB.B]: 30,
     // $ExpectError
-    [$enum.handleNull]: -1,
+    [$enum.handleNull]: -1
+});
+
+// Unnecessary undefined handler causes error
+$enum.mapValue(rgb).with({
+    [RGB.R]: 10,
+    [RGB.G]: 20,
+    [RGB.B]: 30,
+    // $ExpectError
     [$enum.handleUndefined]: -1
+});
+
+// Test enum value computed property names (no compiler error).
+// (only supported as of TS 2.6.1)
+$enum.mapValue(rgb).with({
+    [RGB.R]: 10,
+    [RGB.G]: 20,
+    [RGB.B]: 30
 });
