+++ type_tests/v2_9_plus/mapValue/string-literal-null.dtslint.ts
@@ -1,91 +1,86 @@
-import { $enum } from "../../../dist/types";
+import { $enum } from "ts-enum-util";
 
-enum RGB {
-    R = "r",
-    G = "g",
-    B = "b"
-}
+type RGB = "r" | "g" | "b";
 
-declare const rgb: RGB;
+declare const rgb: RGB | null;
 
 // Return type is inferred
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.G]: 20,
-    [RGB.B]: 30
+    r: 10,
+    g: 20,
+    b: 30,
+    [$enum.handleNull]: -1
 });
 // $ExpectType string
 $enum.mapValue(rgb).with({
-    [RGB.R]: "10",
-    [RGB.G]: "20",
-    [RGB.B]: "30"
+    r: "10",
+    g: "20",
+    b: "30",
+    [$enum.handleNull]: "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.G]: $enum.unhandled,
-    [RGB.B]: 30
+    r: 10,
+    g: $enum.unhandled,
+    b: 30,
+    [$enum.handleNull]: -1
 });
 
 // handleUnexpected is allowed
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.G]: 20,
-    [RGB.B]: 30,
+    r: 10,
+    g: 20,
+    b: 30,
+    [$enum.handleNull]: -1,
     [$enum.handleUnexpected]: -1
 });
 
 // special handlers can be unhandled
 // $ExpectType number
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.G]: 20,
-    [RGB.B]: 30,
+    r: 10,
+    g: 20,
+    b: 30,
+    [$enum.handleNull]: $enum.unhandled,
     [$enum.handleUnexpected]: $enum.unhandled
 });
 
 // Missing value handler causes error
 // $ExpectError
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.B]: 30
+    r: 10,
+    b: 30,
+    [$enum.handleNull]: -1
 });
 
 // Unexpected value handler causes error
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
+    r: 10,
     // $ExpectError
     oops: 42,
-    [RGB.G]: 20,
-    [RGB.B]: 30
+    g: 20,
+    b: 30,
+    [$enum.handleNull]: -1
 });
 
-// Unnecessary null handler causes error
+// missing null handler causes error
+// $ExpectError
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.G]: 20,
-    [RGB.B]: 30,
-    // $ExpectError
-    [$enum.handleNull]: -1
+    r: 10,
+    g: 20,
+    b: 30
 });
 
 // Unnecessary undefined handler causes error
 $enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.G]: 20,
-    [RGB.B]: 30,
+    r: 10,
+    g: 20,
+    b: 30,
+    [$enum.handleNull]: -1,
     // $ExpectError
     [$enum.handleUndefined]: -1
-});
-
-// Test enum value computed property names (no compiler error).
-// (only supported as of TS 2.6.1)
-$enum.mapValue(rgb).with({
-    [RGB.R]: 10,
-    [RGB.G]: 20,
-    [RGB.B]: 30
 });
