+++ type_tests/v2_9_plus/tsconfig.json
@@ -19,6 +19,9 @@
 
         /* Module Resolution Options */
         "moduleResolution": "node",
-        "baseUrl": "./"
+        "baseUrl": "./",
+        "paths": {
+            "ts-enum-util": ["../../dist/types"]
+        }
     }
 }
