+++ type_tests/v2_9_plus/visitValue/string-enum-null.dtslint.ts
@@ -1,12 +1,12 @@
 import { $enum } from "ts-enum-util";
 
 enum RGB {
-    R,
-    G,
-    B
+    R = "r",
+    G = "g",
+    B = "b"
 }
 
-declare const rgb: RGB | null | undefined;
+declare const rgb: RGB | null;
 
 // Test param types
 $enum.visitValue(rgb).with({
@@ -26,12 +26,8 @@ $enum.visitValue(rgb).with({
         // $ExpectType null
         value;
     },
-    [$enum.handleUndefined]: (value) => {
-        // $ExpectType undefined
-        value;
-    },
     [$enum.handleUnexpected]: (value) => {
-        // $ExpectType number
+        // $ExpectType string | undefined
         value;
     }
 });
@@ -41,8 +37,7 @@ $enum.visitValue(rgb).with({
     [RGB.R]: (value) => {},
     [RGB.G]: (value) => {},
     [RGB.B]: (value) => {},
-    [$enum.handleNull]: (value) => {},
-    [$enum.handleUndefined]: (value) => {}
+    [$enum.handleNull]: (value) => {}
 });
 
 // Return type is inferred
@@ -51,16 +46,14 @@ $enum.visitValue(rgb).with({
     [RGB.R]: (value) => 10,
     [RGB.G]: (value) => 20,
     [RGB.B]: (value) => 30,
-    [$enum.handleNull]: (value) => -1,
-    [$enum.handleUndefined]: (value) => -1
+    [$enum.handleNull]: (value) => -1
 });
 // $ExpectType string
 $enum.visitValue(rgb).with({
     [RGB.R]: (value) => "10",
     [RGB.G]: (value) => "20",
     [RGB.B]: (value) => "30",
-    [$enum.handleNull]: (value) => "-1",
-    [$enum.handleUndefined]: (value) => "-1"
+    [$enum.handleNull]: (value) => "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
@@ -69,8 +62,7 @@ $enum.visitValue(rgb).with({
     [RGB.R]: (value) => 10,
     [RGB.G]: $enum.unhandled,
     [RGB.B]: (value) => 30,
-    [$enum.handleNull]: (value) => -1,
-    [$enum.handleUndefined]: (value) => -1
+    [$enum.handleNull]: (value) => -1
 });
 
 // special handlers can be unhandled
@@ -80,7 +72,6 @@ $enum.visitValue(rgb).with({
     [RGB.G]: (value) => 20,
     [RGB.B]: (value) => 30,
     [$enum.handleNull]: $enum.unhandled,
-    [$enum.handleUndefined]: $enum.unhandled,
     [$enum.handleUnexpected]: $enum.unhandled
 });
 
@@ -89,8 +80,7 @@ $enum.visitValue(rgb).with({
 $enum.visitValue(rgb).with<void>({
     [RGB.R]: (value) => {},
     [RGB.B]: (value) => {},
-    [$enum.handleNull]: (value) => {},
-    [$enum.handleUndefined]: (value) => {}
+    [$enum.handleNull]: (value) => {}
 });
 
 // Missing null handler causes error
@@ -98,17 +88,7 @@ $enum.visitValue(rgb).with<void>({
 $enum.visitValue(rgb).with<void>({
     [RGB.R]: (value) => {},
     [RGB.G]: (value) => {},
-    [RGB.B]: (value) => {},
-    [$enum.handleUndefined]: (value) => {}
-});
-
-// Missing undefined handler causes error
-// $ExpectError
-$enum.visitValue(rgb).with<void>({
-    [RGB.R]: (value) => {},
-    [RGB.G]: (value) => {},
-    [RGB.B]: (value) => {},
-    [$enum.handleNull]: (value) => {}
+    [RGB.B]: (value) => {}
 });
 
 // Unexpected value handler causes error
@@ -118,6 +98,15 @@ $enum.visitValue(rgb).with<void>({
     oops: (value) => {},
     [RGB.G]: (value) => {},
     [RGB.B]: (value) => {},
+    [$enum.handleNull]: (value) => {}
+});
+
+// Unnecessary undefined handler causes error
+$enum.visitValue(rgb).with<void>({
+    [RGB.R]: (value) => {},
+    [RGB.G]: (value) => {},
+    [RGB.B]: (value) => {},
     [$enum.handleNull]: (value) => {},
+    // $ExpectError
     [$enum.handleUndefined]: (value) => {}
 });
