+++ type_tests/v2_9_plus/visitValue/string-enum-undefined.dtslint.ts
@@ -1,4 +1,4 @@
-import { $enum } from "../../../dist/types";
+import { $enum } from "ts-enum-util";
 
 enum RGB {
     R = "r",
@@ -6,7 +6,7 @@ enum RGB {
     B = "b"
 }
 
-declare const rgb: RGB | null;
+declare const rgb: RGB | undefined;
 
 // Test param types
 $enum.visitValue(rgb).with({
@@ -22,12 +22,12 @@ $enum.visitValue(rgb).with({
         // $ExpectType RGB.B
         value;
     },
-    [$enum.handleNull]: (value) => {
-        // $ExpectType null
+    [$enum.handleUndefined]: (value) => {
+        // $ExpectType undefined
         value;
     },
     [$enum.handleUnexpected]: (value) => {
-        // $ExpectType string | undefined
+        // $ExpectType string | null
         value;
     }
 });
@@ -37,7 +37,7 @@ $enum.visitValue(rgb).with({
     [RGB.R]: (value) => {},
     [RGB.G]: (value) => {},
     [RGB.B]: (value) => {},
-    [$enum.handleNull]: (value) => {}
+    [$enum.handleUndefined]: (value) => {}
 });
 
 // Return type is inferred
@@ -46,14 +46,14 @@ $enum.visitValue(rgb).with({
     [RGB.R]: (value) => 10,
     [RGB.G]: (value) => 20,
     [RGB.B]: (value) => 30,
-    [$enum.handleNull]: (value) => -1
+    [$enum.handleUndefined]: (value) => -1
 });
 // $ExpectType string
 $enum.visitValue(rgb).with({
     [RGB.R]: (value) => "10",
     [RGB.G]: (value) => "20",
     [RGB.B]: (value) => "30",
-    [$enum.handleNull]: (value) => "-1"
+    [$enum.handleUndefined]: (value) => "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
@@ -62,7 +62,7 @@ $enum.visitValue(rgb).with({
     [RGB.R]: (value) => 10,
     [RGB.G]: $enum.unhandled,
     [RGB.B]: (value) => 30,
-    [$enum.handleNull]: (value) => -1
+    [$enum.handleUndefined]: (value) => -1
 });
 
 // special handlers can be unhandled
@@ -71,7 +71,7 @@ $enum.visitValue(rgb).with({
     [RGB.R]: (value) => 10,
     [RGB.G]: (value) => 20,
     [RGB.B]: (value) => 30,
-    [$enum.handleNull]: $enum.unhandled,
+    [$enum.handleUndefined]: $enum.unhandled,
     [$enum.handleUnexpected]: $enum.unhandled
 });
 
@@ -80,10 +80,10 @@ $enum.visitValue(rgb).with({
 $enum.visitValue(rgb).with<void>({
     [RGB.R]: (value) => {},
     [RGB.B]: (value) => {},
-    [$enum.handleNull]: (value) => {}
+    [$enum.handleUndefined]: (value) => {}
 });
 
-// Missing null handler causes error
+// Missing undefined handler causes error
 // $ExpectError
 $enum.visitValue(rgb).with<void>({
     [RGB.R]: (value) => {},
@@ -98,15 +98,15 @@ $enum.visitValue(rgb).with<void>({
     oops: (value) => {},
     [RGB.G]: (value) => {},
     [RGB.B]: (value) => {},
-    [$enum.handleNull]: (value) => {}
+    [$enum.handleUndefined]: (value) => {}
 });
 
-// Unnecessary undefined handler causes error
+// Unnecessary null handler causes error
 $enum.visitValue(rgb).with<void>({
     [RGB.R]: (value) => {},
     [RGB.G]: (value) => {},
     [RGB.B]: (value) => {},
-    [$enum.handleNull]: (value) => {},
     // $ExpectError
+    [$enum.handleNull]: (value) => {},
     [$enum.handleUndefined]: (value) => {}
 });
